Fix photo triggers - attempt 2 (#47)

+++ init.sql
@@ -228,7 +228,12 @@ CREATE TABLE IF NOT EXISTS photo (
     PRIMARY KEY (person_id, position)
 );
 
+-- TODO: Delete
 CREATE TABLE IF NOT EXISTS photo_graveyard (
+    uuid TEXT PRIMARY KEY
+);
+
+CREATE TABLE IF NOT EXISTS undeleted_photo (
     uuid TEXT PRIMARY KEY
 );
 
@@ -1057,8 +1062,30 @@ RETURNS TABLE(trait_id SMALLINT, ratio FLOAT4) AS $$
 $$ LANGUAGE sql IMMUTABLE LEAKPROOF PARALLEL SAFE;
 
 --------------------------------------------------------------------------------
--- TRIGGERS
+-- TRIGGER - insert_update_search_tables
 --------------------------------------------------------------------------------
+
+CREATE OR REPLACE FUNCTION visible_in_search(p person)
+RETURNS BOOLEAN AS $$
+BEGIN
+    IF p IS NULL THEN
+        RETURN FALSE;
+    END IF;
+
+    RETURN p.activated;
+END;
+$$ LANGUAGE plpgsql;
+
+CREATE OR REPLACE FUNCTION visible_in_quiz_search(p person)
+RETURNS BOOLEAN AS $$
+BEGIN
+    IF p IS NULL THEN
+        RETURN FALSE;
+    END IF;
+
+    RETURN p.activated AND p.has_profile_picture_id = 1;
+END;
+$$ LANGUAGE plpgsql;
 
 CREATE OR REPLACE FUNCTION insert_update_search_tables()
 RETURNS TRIGGER AS $$
@@ -1066,22 +1093,11 @@ DECLARE
     visible_in_search BOOLEAN;
     visible_in_quiz_search BOOLEAN;
 BEGIN
-    visible_in_search := NEW.activated;
-    visible_in_quiz_search := (
-        NEW.has_profile_picture_id = 1 AND
-        NEW.hide_me_from_strangers = FALSE);
+    visible_in_search := visible_in_search(NEW);
+    visible_in_quiz_search := visible_in_quiz_search(NEW);
 
-    IF (visible_in_search = TRUE) THEN
-        IF (visible_in_quiz_search) THEN
-            INSERT INTO search_for_quiz_prospects (person_id, coordinates, personality)
-            VALUES (NEW.id, NEW.coordinates, NEW.personality)
-            ON CONFLICT (person_id)
-            DO UPDATE SET
-                coordinates = EXCLUDED.coordinates,
-                personality = EXCLUDED.personality;
-        END IF;
-
-        INSERT INTO search_for_standard_prospects (person_id, coordinates, personality)
+    IF visible_in_quiz_search THEN
+        INSERT INTO search_for_quiz_prospects (person_id, coordinates, personality)
         VALUES (NEW.id, NEW.coordinates, NEW.personality)
         ON CONFLICT (person_id)
         DO UPDATE SET
@@ -1090,7 +1106,17 @@ BEGIN
     ELSE
         DELETE FROM search_for_quiz_prospects
         WHERE person_id = NEW.id;
+    END IF;
 
+
+    IF visible_in_search THEN
+        INSERT INTO search_for_standard_prospects (person_id, coordinates, personality)
+        VALUES (NEW.id, NEW.coordinates, NEW.personality)
+        ON CONFLICT (person_id)
+        DO UPDATE SET
+            coordinates = EXCLUDED.coordinates,
+            personality = EXCLUDED.personality;
+    ELSE
         DELETE FROM search_for_standard_prospects
         WHERE person_id = NEW.id;
     END IF;
@@ -1105,8 +1131,35 @@ ON person
 FOR EACH ROW
 EXECUTE FUNCTION insert_update_search_tables();
 
+--------------------------------------------------------------------------------
+-- TRIGGER - undeleted_photo
+--------------------------------------------------------------------------------
+
+CREATE OR REPLACE FUNCTION trigger_fn_insert_into_undeleted_photo()
+RETURNS TRIGGER AS $$
+BEGIN
+    INSERT INTO undeleted_photo (uuid)
+    VALUES (NEW.uuid)
+    ON CONFLICT DO NOTHING;
+
+    RETURN NEW;
+END;
+$$ LANGUAGE plpgsql;
+
+CREATE OR REPLACE TRIGGER trigger_fn_insert_into_undeleted_photo_on_insert_photo
+AFTER INSERT OR UPDATE ON photo
+FOR EACH ROW
+EXECUTE FUNCTION trigger_fn_insert_into_undeleted_photo();
+
+CREATE OR REPLACE TRIGGER trigger_fn_insert_into_undeleted_photo_on_insert_onboardee_photo
+AFTER INSERT OR UPDATE ON onboardee_photo
+FOR EACH ROW
+EXECUTE FUNCTION trigger_fn_insert_into_undeleted_photo();
 
 
+--------------------------------------------------------------------------------
+-- TODO: Delete these
+--------------------------------------------------------------------------------
 CREATE OR REPLACE FUNCTION on_delete_photo()
 RETURNS TRIGGER AS $$
 BEGIN
@@ -1129,27 +1182,22 @@ BEFORE DELETE ON photo
 FOR EACH ROW
 EXECUTE FUNCTION on_delete_photo();
 
--- Trigger for DELETE operation on onboardee_photo
 CREATE OR REPLACE TRIGGER trigger_onboardee_photo_delete
 BEFORE DELETE ON onboardee_photo
 FOR EACH ROW
 EXECUTE FUNCTION on_delete_photo();
 
--- Trigger for UPDATE operation
 CREATE OR REPLACE TRIGGER trigger_photo_update
 BEFORE UPDATE ON photo
 FOR EACH ROW
 WHEN (OLD.uuid IS DISTINCT FROM NEW.uuid)
 EXECUTE FUNCTION on_update_photo();
 
--- Trigger for UPDATE operation on onboardee_photo
 CREATE OR REPLACE TRIGGER trigger_onboardee_photo_update
 BEFORE UPDATE ON onboardee_photo
 FOR EACH ROW
 WHEN (OLD.uuid IS DISTINCT FROM NEW.uuid)
 EXECUTE FUNCTION on_update_photo();
-
-
 
 CREATE OR REPLACE FUNCTION remove_from_photo_graveyard()
 RETURNS TRIGGER AS $$
@@ -1213,6 +1261,126 @@ CREATE OR REPLACE TRIGGER trigger_refresh_has_profile_picture_id
 AFTER INSERT OR DELETE ON photo
 FOR EACH ROW
 EXECUTE FUNCTION trigger_fn_refresh_has_profile_picture_id();
+
+--------------------------------------------------------------------------------
+-- MIGRATION
+-- TODO: Delete after migration is complete
+--------------------------------------------------------------------------------
+
+CREATE OR REPLACE FUNCTION refresh_search_tables(p_person_id INT)
+RETURNS INTEGER AS $$
+    WITH selected_person AS (
+        SELECT
+            id,
+            coordinates,
+            personality,
+            activated,
+            has_profile_picture_id,
+            hide_me_from_strangers
+        FROM
+            person
+        WHERE
+            id = p_person_id
+    ), visible_in_standard_search AS (
+        SELECT (
+            activated
+        ) AS x
+        FROM selected_person
+    ), visible_in_quiz_search AS (
+        SELECT (
+            activated AND
+            NOT hide_me_from_strangers AND
+            has_profile_picture_id = 1
+        ) AS x
+        FROM selected_person
+    ), inserted_standard_search AS (
+        INSERT INTO search_for_standard_prospects (
+            person_id, coordinates, personality
+        )
+        SELECT
+            id AS person_id, coordinates, personality
+        FROM
+            selected_person
+        WHERE
+            (SELECT x FROM visible_in_standard_search)
+        ON CONFLICT (person_id) DO UPDATE SET
+            coordinates = EXCLUDED.coordinates,
+            personality = EXCLUDED.personality
+        RETURNING *
+    ), deleted_standard_search AS (
+        DELETE FROM search_for_standard_prospects
+        USING selected_person
+        WHERE person_id = selected_person.id AND
+            (SELECT NOT x FROM visible_in_standard_search)
+        RETURNING *
+    ), inserted_quiz_search AS (
+        INSERT INTO search_for_quiz_prospects (
+            person_id, coordinates, personality
+        )
+        SELECT
+            id AS person_id, coordinates, personality
+        FROM
+            selected_person
+        WHERE
+            (SELECT x FROM visible_in_quiz_search)
+        ON CONFLICT (person_id) DO UPDATE SET
+            coordinates = EXCLUDED.coordinates,
+            personality = EXCLUDED.personality
+        RETURNING *
+    ), deleted_quiz_search AS (
+        DELETE FROM search_for_quiz_prospects
+        USING selected_person
+        WHERE person_id = selected_person.id AND
+            (SELECT NOT x FROM visible_in_quiz_search)
+        RETURNING *
+    )
+    SELECT
+        (SELECT COUNT(*) from inserted_standard_search) +
+        (SELECT COUNT(*) from deleted_standard_search) +
+        (SELECT COUNT(*) from inserted_quiz_search) +
+        (SELECT COUNT(*) from deleted_quiz_search)
+$$ LANGUAGE sql;
+
+-- Drop the triggers
+DROP TRIGGER trigger_photo_delete ON photo;
+DROP TRIGGER trigger_onboardee_photo_delete ON onboardee_photo;
+DROP TRIGGER trigger_photo_update ON photo;
+DROP TRIGGER trigger_onboardee_photo_update ON onboardee_photo;
+DROP TRIGGER trigger_photo_insert ON photo;
+DROP TRIGGER trigger_onboardee_photo_insert ON onboardee_photo;
+
+-- Drop the functions
+DROP FUNCTION on_delete_photo();
+DROP FUNCTION on_update_photo();
+DROP FUNCTION remove_from_photo_graveyard();
+
+
+CREATE OR REPLACE FUNCTION assert_table_non_empty(tbl_name text) RETURNS void AS $$
+DECLARE
+    row_count INTEGER;
+BEGIN
+    EXECUTE format('SELECT count(*) FROM %I', tbl_name) INTO row_count;
+
+    IF row_count = 0 THEN
+        RAISE EXCEPTION 'Table % is empty!', tbl_name;
+    END IF;
+
+    RAISE NOTICE 'Table % has % rows', tbl_name, row_count;
+END;
+$$ LANGUAGE plpgsql;
+
+INSERT INTO photo_graveyard (uuid) VALUES ('dummy');
+
+INSERT INTO undeleted_photo (uuid)
+SELECT uuid
+FROM photo_graveyard
+ON CONFLICT DO NOTHING;
+
+SELECT assert_table_non_empty('undeleted_photo');
+
+DROP TABLE photo_graveyard;
+
+SELECT id, refresh_search_tables(id) FROM person;
 
 --------------------------------------------------------------------------------
